app/requirements.txt
transformers
fastapi
uvicorn
torch
streamlit
python-dotenv
requests